fix: missing check for moderators

+++ src/channels/pf-bugpin.js
@@ -1,6 +1,6 @@
 module.exports = {
   id: "1019668981457625189",
   execute(client, message) {
-    message.delete();
+    if (!message.author.moderator) message.delete();
   }
 }
