+++ src/channels/pf-dcpin.js
@@ -1,6 +1,6 @@
 module.exports = {
   id: "1019683152572190780",
   execute(client, message) {
-    message.delete();
+    if (!message.author.moderator) message.delete();
   }
 }
